smart-netfix/requirements.txt
fastapi
uvicorn
speedtest-cli
pydantic
python-multipart
streamlit
requests
fpdf2
pytz
tzlocal
